refactor: move render to separated file
